+++ services/LotideService/Instance.ts
@@ -1,0 +1,9 @@
+import { lotideRequest } from "./util";
+
+export async function getInstanceInfo(
+  ctx: LotideContext,
+): Promise<InstanceInfo> {
+  return lotideRequest(ctx, "GET", "instance", undefined, true).then(data =>
+    data.json(),
+  );
+}
