+++ services/LotideService/Notification.ts
@@ -1,0 +1,9 @@
+import { lotideRequest } from "./util";
+
+export async function getNotifications(
+  ctx: LotideContext,
+): Promise<UserNotification[]> {
+  return lotideRequest(ctx, "GET", "users/~me/notifications").then(data =>
+    data.json(),
+  );
+}
