+++ services/LotideService/index.ts
@@ -1,0 +1,8 @@
+export * from "./Community";
+export * from "./Instance";
+export * from "./Notification";
+export * from "./Post";
+export * from "./Reply";
+export * from "./User";
+
+export * from "./util";
